refactor: level guide page
